Add document-collections.
A document-collection is just a collection of documents. It could either
be created when uploading docs to create a new chat, or in future also
when connecting a new data external source. And it maps to a collection
in vector db. A chat has either zero or one associated
document-collection.

This was required because chat-ids (nanoids) can't be used as collection
name in chroma db (contain invalid characters). So we have to keep track
of vector-db collection name ourselves.

Introducing document collection in our data model allows us to do more
with it in the future. For example, we can use document collections when
building connect-with-external-data-source and make indexing an offline
step. This will then allow the users to connect with a data source once,
wait for it to be indexed, and then create as many chats with it as
needed.