Publish message on document upload.
The document upload endpoint now publishes a message to the indexing
queue if indexing-mode is offline. FE passes offline when creating
document collection, and online when creating chat.

This also cleans up the code: Moved AMQP configs to env vars, and
converted task-masker to use async/await instead of callbacks!