Make task-master process one document at a time. (#45)
1. set noAck to false. This is because we need the unprocessed messages
to go back to the queue. We explicitly ack the message once the message
is processed. Current timeout for ack is 30 mins by default. For now, it
should be fine because we expect the message to get processed within 30
mins. If it goes over it, task master will stop.
2. consume one message at a time. In future, we should play around with
this config when we add more consumers

Co-authored-by: qchroma <qchroma>